Clustering TS: cleanup descriptor disable test
